refactor: move loadData init-time flow wiring into DepositDataLoader (#4999)
Move the loadData init-time wiring (option-list build + initial state
seed, coin-list derivation, loadAddress call, and the three
sequence-sensitive flow collectors) out of DepositFormViewModel into a
single DepositDataLoader.wireInitialState entry point.

loadData stays a thin VM entry point: it stores params, calls
collectAmountChanges(), then forwards into the loader. viewModelScope
ownership stays in the VM (passed in as the assisted scope); collector
order/structure, distinctUntilChanged operators, and the SecuredAsset
address population inside the collect are preserved bit-identically. No
public ViewModel API change.
