+++ app/src/main/java/com/vultisig/wallet/ui/models/deposit/DepositFormViewModel.kt
@@ -7,7 +7,6 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.vultisig.wallet.R
 import com.vultisig.wallet.data.api.models.thorchain.RujiStakeBalances
-import com.vultisig.wallet.data.crypto.ThorChainHelper.Companion.SECURE_ASSETS_TICKERS
 import com.vultisig.wallet.data.models.Account
 import com.vultisig.wallet.data.models.Address
 import com.vultisig.wallet.data.models.AddressBookEntry
@@ -48,7 +47,6 @@ import com.vultisig.wallet.ui.navigation.Navigator
 import com.vultisig.wallet.ui.navigation.Route
 import com.vultisig.wallet.ui.navigation.SendDst
 import com.vultisig.wallet.ui.screens.select.AssetSelected
-import com.vultisig.wallet.ui.screens.v2.defi.model.DeFiNavActions
 import com.vultisig.wallet.ui.utils.UiText
 import com.vultisig.wallet.ui.utils.asUiText
 import com.vultisig.wallet.ui.utils.textAsFlow
@@ -67,11 +65,8 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.firstOrNull
-import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
@@ -379,177 +374,26 @@ constructor(
 
         collectAmountChanges()
 
-        val depositOptions =
-            when (chain) {
-                Chain.ThorChain ->
-                    listOf(
-                        DepositOption.Bond,
-                        DepositOption.Unbond,
-                        DepositOption.Leave,
-                        DepositOption.Custom,
-                        DepositOption.Merge,
-                        DepositOption.UnMerge,
-                        DepositOption.WithdrawSecuredAsset,
-                    )
-
-                Chain.MayaChain -> listOf(DepositOption.Leave, DepositOption.Custom)
-
-                Chain.Kujira,
-                Chain.Osmosis -> listOf(DepositOption.TransferIbc)
-
-                Chain.GaiaChain -> listOf(DepositOption.TransferIbc, DepositOption.Switch)
-                Chain.Ton -> {
-                    listOf(DepositOption.Stake, DepositOption.Unstake)
-                }
-                else ->
-                    buildList {
-                        //                    add(DepositOption.Stake)
-                        //                    add(DepositOption.Unstake)
-                        if (chain.ticker() in SECURE_ASSETS_TICKERS) add(DepositOption.SecuredAsset)
-                    }
-            }
-        val depositOption = depositOptions.first()
-        val defaultToken =
-            when (chain) {
-                Chain.MayaChain -> Coins.MayaChain.CACAO
-                else -> Coins.ThorChain.RUNE
-            }
-        _state.update {
-            it.copy(
-                depositMessage = R.string.deposit_message_deposit_title.asUiText(chain.raw),
-                depositOptions = depositOptions,
-                depositOption = depositOption,
-                depositChain = chain,
-                selectedToken = defaultToken,
-            )
-        }
-
-        val coinList =
-            tokensToMerge.let {
-                if (chain == Chain.Osmosis) it.filter { it.ticker.equals("LVN", ignoreCase = true) }
-                else it
-            }
-        _state.update {
-            it.copy(
-                selectedCoin = coinList.first(),
-                coinList = coinList,
-                selectedUnMergeCoin = coinList.first(),
-            )
-        }
-
-        dataLoader.loadAddress(vaultId, chain)
-
-        address
-            .filterNotNull()
-            .onEach { address ->
-                val selectedToken = address.accounts.find { it.token.isNativeToken }?.token
-                selectedToken?.let { _state.update { it.copy(selectedToken = selectedToken) } }
-                if (depositTypeAction == DeFiNavActions.ADD_LP.type) {
-                    gasFeeHelper.loadGasFeeForDisplay(
-                        scope = viewModelScope,
-                        vaultId = vaultId,
-                        chain = chain,
-                        address = address,
-                        onResult = { totalGas, estimatedFee ->
-                            _state.update {
-                                it.copy(totalGas = totalGas, estimatedFee = estimatedFee)
-                            }
-                        },
-                    )
-                }
-            }
-            .launchIn(viewModelScope)
-
-        viewModelScope.launch {
-            combine(
-                    state.map { it.selectedCoin }.distinctUntilChanged(),
-                    address.filterNotNull(),
-                    state.map { it.depositOption }.distinctUntilChanged(),
-                    state.map { it.selectedToken }.distinctUntilChanged(),
-                ) { selectedMergeToken, address, depositOption, selectedToken ->
-                    var targetTicker: String?
-
-                    val account =
-                        when (depositOption) {
-                            DepositOption.Switch,
-                            DepositOption.TransferIbc,
-                            DepositOption.Merge -> {
-                                targetTicker = selectedMergeToken.ticker
-                                address.accounts.find {
-                                    it.token.ticker.equals(
-                                        selectedMergeToken.ticker,
-                                        ignoreCase = true,
-                                    )
-                                }
-                            }
-
-                            DepositOption.Custom -> {
-                                targetTicker = selectedToken.ticker
-                                address.accounts.find { it.token.id == selectedToken.id }
-                            }
-
-                            else -> {
-                                val account = address.accounts.find { it.token.isNativeToken }
-                                targetTicker = account?.token?.ticker
-                                account
-                            }
-                        }
-
-                    if (depositOption != DepositOption.RemoveLiquidity) {
-                        updateTokenAmount(account, chain, targetTicker, vaultId)
-                    }
-
-                    depositOption
-                }
-                .collect { depositOption ->
-                    // Populate the user's own THORChain address for the SecuredAsset form here,
-                    // outside the (pure) transform, and never resolve the inbound vault as a side
-                    // effect — that is done synchronously in SecuredAssetStrategy so the
-                    // destination always matches the currently-selected asset's chain.
-                    if (depositOption == DepositOption.SecuredAsset) {
-                        securedAssetLoader.collectSecuredAssetAddresses()
-                    }
-                    dataLoader.setMetadataInfo()
-                }
-        }
-
-        viewModelScope.launch {
-            combine(
-                    state.map { it.selectedCoin }.distinctUntilChanged(),
-                    state.map { it.depositOption }.distinctUntilChanged(),
-                ) { selectedMergeToken, depositOption ->
-                    when (depositOption) {
-                        DepositOption.TransferIbc,
-                        DepositOption.Switch -> {
-                            // special case, because of all supported merge tokens only lvn is
-                            // osmosis native
-                            val dstChainList =
-                                if (selectedMergeToken.ticker == "LVN") {
-                                    when (chain) {
-                                        Chain.Osmosis -> listOf(Chain.GaiaChain)
-                                        else -> listOf(Chain.Osmosis)
-                                    }
-                                } else {
-                                    listOf(
-                                            Chain.GaiaChain,
-                                            Chain.Kujira,
-                                            Chain.Osmosis,
-                                            Chain.Noble,
-                                            Chain.Akash,
-                                        )
-                                        .filter { it != chain }
-                                }
-
-                            _state.update { it.copy(dstChainList = dstChainList) }
-
-                            selectDstChain(dstChainList.first())
-                        }
-
-                        else -> Unit
-                    }
-                }
-                .collect {}
-        }
+        dataLoader.wireInitialState(
+            vaultId = vaultId,
+            chain = chain,
+            tokensToMerge = tokensToMerge,
+            state = _state,
+            updateTokenAmount = ::updateTokenAmount,
+            selectDstChain = ::selectDstChain,
+            collectSecuredAssetAddresses = securedAssetLoader::collectSecuredAssetAddresses,
+            loadGasFeeForDisplay = { vaultId, chain, address ->
+                gasFeeHelper.loadGasFeeForDisplay(
+                    scope = viewModelScope,
+                    vaultId = vaultId,
+                    chain = chain,
+                    address = address,
+                    onResult = { totalGas, estimatedFee ->
+                        _state.update { it.copy(totalGas = totalGas, estimatedFee = estimatedFee) }
+                    },
+                )
+            },
+        )
     }
 
     fun selectBondAsset(asset: String) {
